requirements.txt
pytesseract
Pillow
PyMuPDF
python-docx
requests
python-dotenv
PyYAML
